backend/requirements.txt
Django==3.2.3
pillow==10.0.0
djangorestframework==3.14.0